+++ FrontEnd/frontend/src/App.js
@@ -12,8 +12,8 @@ import { Footer } from './components/Footer';
 import { Producto } from './components/producto/Producto';
 import { CarritoProvider } from './components/carrito/CarritoContext';
 import { ResetPassword } from "./components/login/ResetPassword";
-import {InicioAdmin} from "./components/inicioAdmin/InicioAdmin";
-import {InicioClientes} from "./components/inicioAdmin/inicioClientes";
+import { InicioAdmin } from "./components/inicioAdmin/InicioAdmin";
+import { InicioAdminPrueba } from "./components/inicioAdmin/InicioAdminPrueba";
 // import { CargarProducto } from "./components/cargarProducto/CargarProducto";
 
 function App() {
@@ -31,7 +31,7 @@ function App() {
               <Route path="/info" element={<Info />} />
               <Route path='/productos/:categoria' element={<Producto />}/>
               <Route path='/productos' element={<Producto />}/>
-              <Route path="/inicioClientes" element={<InicioClientes />} />
+              <Route path="/inicioAdminPrueba" element={<InicioAdminPrueba />} />
               <Route path="/inicioAdmin" element={<InicioAdmin />} />
               <Route path="/principal" element={<Principal />} />
               <Route path='/' element={<Principal />} />
