+++ FrontEnd/frontend/src/components/footer.css
@@ -1,3 +1,4 @@
+
 .footer {
     background-color: #FCBB3A !important;
     color: white;
